pyproject.toml
[tool.poetry]
name = "wisper-test"
version = "0.1.0"
description = ""
authors = ["douqiting <229813475@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llvmlite = "^0.44.0"
numba = "^0.61.0"
openai-whisper = "^20240930"
numpy = "1.24.0"
pyaudio = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
